requirements.txt
bcrypt==3.2.0
boto3==1.19.12
celery==5.1.2
django-axes==5.26.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-filter==21.1
django-storages==1.12.3
django-templated-email==3.0.0
Django==3.2.9
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
markdown==3.3.4
openpyxl==3.0.9
psycopg2-binary==2.9.1
pygments==2.10.0
pytest-django==4.4.0
pytest-flake8==1.0.7
pytest==6.2.5
python-magic==0.4.24
pyyaml==6.0
redis==3.5.3
requests==2.26.0
sentry-sdk==1.4.3
uritemplate==4.1.1

